+++ packages/lib/factorio/server.js
@@ -13,6 +13,22 @@ const libErrors = require("@clusterio/lib/errors");
 
 
 /**
+ * Escapes text for inclusion in a RegExp
+ *
+ * Adds \ character in front of special meta characters in the passsed in
+ * text so that it can be embedded into a RegExp and only match the text.
+ *
+ * See https://stackoverflow.com/a/9310752
+ *
+ * @param {string} text - Text to escape RegExp meta chars in.
+ * @returns {string} escaped text.
+ * @inner
+ */
+function escapeRegExp(text) {
+	return text.replace(/[-[\]{}()*+?.,\\^$|#\s]/g, "\\$&");
+}
+
+/**
  * Determines the version of Factorio the datadir is pointing to by
  * reading the changelog.txt in it.
  *
@@ -302,13 +318,15 @@ function parseOutput(line, source) {
 		// A seconds output has two different kinds of formats: The first is a
 		// log level and source location and has a format of
 		// "Level File.cpp:123: message"
-		const secLogRegex = /^(\w+) (\w+\.cpp:\d+): (.*)$/;
+		// "Script @/path/to/file.lua:123: message"
+		// "Script =(command):123: message"
+		const secLogRegex = /^(\w+) ((\w+\.cpp:\d+)|([@=].*?:\d+)): (.*)$/;
 		let secLogMatch = secLogRegex.exec(secContent);
 		if (secLogMatch) {
 			output.type = "log";
 			output.level = secLogMatch[1];
 			output.file = secLogMatch[2];
-			output.message = secLogMatch[3];
+			output.message = secLogMatch[5];
 
 		// The other possibility is that the content is a generic message
 		} else {
@@ -497,6 +515,9 @@ class FactorioServer extends events.EventEmitter {
 	 * @param {number} options.gamePort - UDP port to host game on.
 	 * @param {number} options.rconPort - TCP port to use for RCON.
 	 * @param {string} options.rconPassword - Password use for RCON.
+	 * @param {boolean} options.enableWhitelist - Turn on whitelisting.
+	 * @param {boolean} options.verboseLogging - Enable verbose logging.
+	 * @param {boolean} options.stripPaths - Strip paths in the console.
 	 */
 	constructor(factorioDir, writeDir, options) {
 		super();
@@ -517,11 +538,26 @@ class FactorioServer extends events.EventEmitter {
 		this.rconPassword = options.rconPassword;
 		/** Enable player whitelist */
 		this.enableWhitelist = options.enableWhitelist;
+		/** Enable verbose logging */
+		this.verboseLogging = options.verboseLogging;
 		this._state = "new";
 		this._server = null;
 		this._rconClient = null;
 		this._rconReady = false;
 		this._gameReady = false;
+
+		if (options.stripPaths) {
+			let chars = new Set(this.writePath("temp"));
+			chars.delete(":"); // Having a colon could lead to matching the line number
+			chars = [...chars].join("");
+
+			this._stripRegExp = new RegExp(
+				`(${escapeRegExp(this.writePath("temp", "currently-playing", path.sep))})|` +
+				`(${escapeRegExp(this.writePath(path.sep))})|` +
+				`(\\.\\.\\.[${escapeRegExp(chars)}]*?currently-playing\\${path.sep})`,
+				"g"
+			);
+		}
 
 		// Array of possible causes for an unexpected shutdown
 		this._unexpected = [];
@@ -606,7 +642,13 @@ class FactorioServer extends events.EventEmitter {
 
 		this.emit(source, line);
 
-		let output = parseOutput(line.toString("utf-8"), source);
+		line = line.toString("utf-8");
+		if (this._stripRegExp) {
+			line = line.replace(this._stripRegExp, "");
+		}
+
+
+		let output = parseOutput(line, source);
 		heuristicLoop: for (let heuristic of outputHeuristics) {
 			for (let [name, expected] of Object.entries(heuristic.filter)) {
 				if (!Object.hasOwnProperty.call(output, name)) {
@@ -774,6 +816,7 @@ class FactorioServer extends events.EventEmitter {
 			[
 				"--config", this.writePath("config.ini"),
 				"--create", this.writePath("saves", name),
+				...(this.verboseLogging ? ["--verbose"] : []),
 			],
 			{
 				detached: true,
@@ -816,6 +859,7 @@ class FactorioServer extends events.EventEmitter {
 				"--rcon-port", this.rconPort,
 				"--rcon-password", this.rconPassword,
 				...(this.enableWhitelist ? ["--use-server-whitelist"] : []),
+				...(this.verboseLogging ? ["--verbose"] : []),
 			],
 			{
 				detached: true,
@@ -850,6 +894,7 @@ class FactorioServer extends events.EventEmitter {
 				"--rcon-port", this.rconPort,
 				"--rcon-password", this.rconPassword,
 				...(this.enableWhitelist ? ["--use-server-whitelist"] : []),
+				...(this.verboseLogging ? ["--verbose"] : []),
 			],
 			{
 				detached: true,