+++ packages/slave/slave.js
@@ -135,8 +135,11 @@ class Instance extends libLink.Link{
 			rconPort: this.config.get("factorio.rcon_port"),
 			rconPassword: this.config.get("factorio.rcon_password"),
 			enableWhitelist: this.config.get("factorio.enable_whitelist"),
+			verboseLogging: this.config.get("factorio.verbose_logging"),
+			stripPaths: this.config.get("factorio.strip_paths"),
 		};
 
+		this._status = "stopped";
 		this._running = false;
 		this.server = new libFactorio.FactorioServer(
 			factorioDir, this._dir, serverOptions
@@ -179,9 +182,21 @@ class Instance extends libLink.Link{
 	}
 
 	notifyStatus(status) {
+		this._status = status;
 		libLink.messages.instanceStatusChanged.send(this, {
 			instance_id: this.config.get("instance.id"), status,
 		});
+	}
+
+	/**
+	 * Current state of the instance
+	 *
+	 * One of stopped, starting, running, creating_save and exporting_data
+	 *
+	 * @returns {string} instance status.
+	 */
+	get status() {
+		return this._status;
 	}
 
 	notifyExit() {
@@ -218,6 +233,7 @@ class Instance extends libLink.Link{
 	}
 
 	async init(pluginInfos) {
+		this.notifyStatus("starting");
 		await this.server.init();
 
 		// load plugins
@@ -794,6 +810,22 @@ class InstanceConnection extends libLink.Link {
 		return await request.send(this.slave, message.data);
 	}
 
+	async forwardRequestToInstance(message, request) {
+		let instanceId = message.data.instance_id;
+		let instanceConnection = this.slave.instanceConnections.get(instanceId);
+		if (!instanceConnection) {
+			// Instance is probably on another slave
+			return await request.send(this.slave, message.data);
+		}
+
+		if (request.plugin && !instanceConnection.plugins.has(request.plugin)) {
+			throw new libErrors.RequestError(`Instance ID ${instanceId} does not have ${request.plugin} plugin loaded`);
+		}
+
+		return await request.send(instanceConnection, message.data);
+	}
+
+
 	async forwardEventToInstance(message, event) {
 		let instanceId = message.data.instance_id;
 		let instanceConnection = this.slave.instanceConnections.get(instanceId);
@@ -1086,7 +1118,6 @@ class Slave extends libLink.Link {
 		if (instanceInfo) {
 			instanceInfo.config.update(serialized_config, true);
 			console.log(`Clusterio | Updated config for ${instanceInfo.path}`);
-			// TODO: Notify of update if instance is running
 
 		} else {
 			let instanceConfig = new libConfig.InstanceConfig();
@@ -1102,8 +1133,17 @@ class Slave extends libLink.Link {
 			};
 			this.instanceInfos.set(instance_id, instanceInfo);
 			console.log(`Clusterio | assigned instance ${instanceConfig.get("instance.name")}`);
+
 		}
 
+		// Somewhat hacky, but in the event of a lost session the status is
+		// resent on assigment since the master server sends an assigment
+		// request for all the instances it knows should be on this slave.
+		let instanceConnection = this.instanceConnections.get(instance_id);
+		libLink.messages.instanceStatusChanged.send(this, {
+			instance_id,
+			status: instanceConnection ? instanceConnection.status : "stopped",
+		});
 
 		// save a copy of the instance config
 		let warnedOutput = {
@@ -1114,9 +1154,6 @@ class Slave extends libLink.Link {
 			path.join(instanceInfo.path, "instance.json"),
 			JSON.stringify(warnedOutput, null, 4)
 		);
-
-		// Notify master the instance is no longer usassigned
-		libLink.messages.instanceStatusChanged.send(this, { instance_id, status: "stopped" });
 	}
 
 	/**
@@ -1140,10 +1177,10 @@ class Slave extends libLink.Link {
 		let instance = new Instance(
 			this, connectionClient, instanceInfo.path, this.config.get("slave.factorio_directory"), instanceInfo.config
 		);
+
+		this.instanceConnections.set(instanceId, instanceConnection);
 		await instance.init(this.pluginInfos);
 
-		// XXX: race condition on multiple simultanious calls
-		this.instanceConnections.set(instanceId, instanceConnection);
 		return instanceConnection;
 	}